Add map to all pages
